Recreate the parser with scopes
Add more semantic meaning of functions and instructions.
This allows e.g. selecting functions or arguments and context-sensitive
highlighting (e.g. for a `%struct` in it's definition or usage as a
type).

The implementation is based on the parser implementation in llvm and the
llvm language reference.
The instruction tests are taken from the llvm language reference.

I ran the parser on the llvm test suite (filtering out the tests that
contain invalid IR) and it doesn't emit ERROR or MISSING on them.